requirements.txt
pypdf
reportlab
requests
google-genai
python-pptx
